Refactor: user 상태 위치 변경
