+++ src/hooks/recoil/user/atom.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
 
-import { UserType } from '../../types';
+import { UserType } from '../../../types';
 
 export const userState = atom<UserType>({
   key: 'userState',
