+++ src/pages/MiniTadak.tsx
@@ -1,6 +1,6 @@
 import { useLocation } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import { userState } from '../hooks/recoil/atom';
+import { userState } from '../hooks/recoil/user/atom';
 import { getMicrophoneAndCameraTracks } from '../agora/config';
 import useAgora from '../hooks/useAgora';
